+++ lsf/lib/liblsf/initenv.c
@@ -178,7 +178,7 @@ readconfenv_ (struct config_param *pList1, struct config_param *pList2, char *co
     FILE *fp;
     char filename[MAXFILENAMELEN];
     struct config_param *plp;
-    uint lineNum = 0;
+    size_t lineNum = 0;
     int saveErrNo = 0;
 
     if (pList1)
